资料下载错误纠正

+++ repository/api/download.php
@@ -16,35 +16,35 @@ session_start();
 $url = "upload_file/".$_GET['url'];
 $filename = $_GET['filename'];
 $srcFile = "../../" . $url;		// 需要下载的文件的真正路径
-$flag = false;
 
-if(is_dir($srcFile)) {
-	// 打包下载文件夹
-	$zip = new zip();
-
-	$list = explode("/", $url);
-	$destFile = $srcFile . ".zip";
-	$zip->compress_dir($srcFile, $destFile);
-	$flag = FileProcess::download($destFile);
-	// 下载完成后删除zip包
-	unlink($destFile);
-} else {
-	// 直接下载文件
-	$flag = FileProcess::download($srcFile);
-}
-
-if($flag !== true) {
-	// 文件不存在，向前台返回一个结果
+if(!file_exists($srcFile)) {
 	echo json_encode(['code' => 0, 'msg' => '文件不存在']);
 	exit;
+} else {
+	if(is_dir($srcFile)) {
+		// 打包下载文件夹
+		$zip = new zip();
+
+		$list = explode("/", $url);
+		$destFile = $srcFile . ".zip";
+		$zip->compress_dir($srcFile, $destFile);
+		FileProcess::download($destFile);
+		// 下载完成后删除zip包
+		unlink($destFile);
+	} else {
+		// 直接下载文件
+		FileProcess::download($srcFile);
+	}
 }
 
-// 如果文件存在并且下载成功，则需要更新数据库
+// 文件存在，先建立下载项
 $db = new Db();
 
 // 建立一条下载记录
 $user = $_SESSION['user'];
-$dlFile = $db->select("download_file", ['email' => $uesr->getEmail(), 'filename' => $filename, 'path' => $list['path']]);
+$list = FileProcess::splitPathAndFilename($url);		// 拆分文件名和路径名
+$dlFile = $db->select("download_file", ['email' => $user->getEmail(), 'filename' => $filename, 'path' => $list['path']]);
+
 if($dlFile === NULL) {
 	// 只有第一次下载资料才能对下载量造成影响
 
@@ -56,8 +56,6 @@ if($dlFile === NULL) {
 	$dlFile->setTime(time());
 	$db->insert("download_file", $dlFile);
 
-	// 拆分文件名和路径名
-	$list = FileProcess::splitPathAndFilename($url);
 	$file = $db->select("file", ['filename' => $filename, 'path' => $list['path']]);
 	if($file !== NULL) {
 		$file->setDownload($file->getDownload() + 1);
@@ -69,6 +67,5 @@ if($dlFile === NULL) {
 	$dlFile->setIsmark(0);
 	$db->update("download_file", $dlFile);
 }
-
 echo json_encode(['code' => 1, 'msg' => '下载成功，请稍后对资料文件进行评价']);
 ?>
